pyproject.toml
[project]
name = "gcyt-dlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.112",
	"uvicorn[standard]>=0.30",
	"yt-dlp>=2024.3",
]

[project.optional-dependencies]
gcs = [
	"google-cloud-storage>=2.17.0",
]
